+++ src/lib/constants/settings/index.ts
@@ -1,6 +1,6 @@
 import { Files, Folder, Monitor, Moon, Sun, Gauge, ChevronsLeftRightEllipsis, ShieldCheck, SearchCode, LucideProps } from "lucide-react";
-import { BackendSettings, IDateFormatSettings, ISettings, IThemeSettings, ScanOptionKeys } from "@/lib/types/settings";
-import { DateFormatType, ScanProfiles, ScanOptionGroup, BehaviorMode, BadgeVisibility } from "@/lib/types/enums";
+import { IBackendSettings, IDateFormatSettings, ISettings, IThemeSettings, ScanOptionKeys } from "@/lib/types/settings";
+import { DateFormatType, ScanProfile, ScanOptionGroup, BehaviorMode, BadgeVisibility } from "@/lib/types/enums";
 
 export const SCAN_OPTION_ICON: Record<ScanOptionGroup,React.ForwardRefExoticComponent<Omit<LucideProps, "ref"> & React.RefAttributes<SVGSVGElement>>> = {
      detection: ShieldCheck,
@@ -18,7 +18,7 @@ export const DEFAULT_SETTINGS: ISettings = {
      confirmStopScan: true,
      autoScrollText: true,
      maxLogLines: 500,
-     currScanProfile: ScanProfiles.Custom,
+     currScanProfile: ScanProfile.Custom,
      realTime: true,
      enableSchedulerUI: true,
      notifOnScanStart: false,
@@ -27,7 +27,7 @@ export const DEFAULT_SETTINGS: ISettings = {
      behavior: BehaviorMode.Balanced,
      badgeVisibility: BadgeVisibility.IconText
 }
-export const DEFAULT_BACKEND_SETTINGS: BackendSettings = {
+export const DEFAULT_BACKEND_SETTINGS: IBackendSettings = {
      scanProfiles: {
           main: {},
           custom: {},