+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ---
 ## [1.0.3] - Unreleased
+### Added
+- Feature to Export / Import Some Settings
+- Translations to every dialog titles
+### Fixed
+- Error toast description byg
 
 ## [1.0.2] - 2026-03-07
 ### Added